New UI and new feature
The trash can icon that is responsible for clearing all filters is now in the sidebar. (On mobile view.)

Added numbers or texts that tell the user what filters are selected to each accordion in the sidebar. (On mobile view)

Added an filter icon to the filter button for mobile view.

The sign beside the state will now display the selected state.

Role filter is now checkbox again, it allows the user to select multiple roles at the same time.

All checkbox are represented by squares and all radios are represented by circles.

Color changes, from indigo to emerald.